+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added title for the connections dropdown list for simplicity.
 - Added automatically renaming the file to a numbered name if file already exists.
+- Added replace warning if the user intentionally renamed the file the same as an existing one.
 
 ### Fixed
 - Fixed clicking "Pause All" and "Resume All" buttons result in crashing.